+++ tests/test_central_pydevccu.py
@@ -29,7 +29,7 @@ class TestCentralPyDevCCU:
         assert central.model == "PyDevCCU"
         assert central.client_coordinator.get_client(interface_id=const.INTERFACE_ID).model == "PyDevCCU"
         assert central.client_coordinator.primary_client.model == "PyDevCCU"
-        assert len(central.device_registry.devices) == 396
+        assert len(central.device_registry.devices) == 397
 
         data = {}
         for device in central.device_registry.devices:
@@ -85,7 +85,7 @@ class TestCentralPyDevCCU:
                 channel_type_names.add(channel.type_name)
 
         channel_type_names = sorted(channel_type_names)
-        assert len(channel_type_names) == 558
+        assert len(channel_type_names) == 559
         ce_channels = {}
         for cdp in custom_dps:
             if cdp.device.model not in ce_channels:
@@ -153,20 +153,20 @@ class TestCentralPyDevCCU:
         for sv in central.hub_coordinator.sysvar_data_points:
             assert hasattr(sv, "__dict__") is False
 
-        assert usage_types[DataPointUsage.CDP_PRIMARY] == 276
+        assert usage_types[DataPointUsage.CDP_PRIMARY] == 279
         assert usage_types[DataPointUsage.CDP_SECONDARY] == 164
-        assert usage_types[DataPointUsage.CDP_VISIBLE] == 155
-        assert usage_types[DataPointUsage.DATA_POINT] == 4050
-        assert usage_types[DataPointUsage.NO_CREATE] == 4329
+        assert usage_types[DataPointUsage.CDP_VISIBLE] == 156
+        assert usage_types[DataPointUsage.DATA_POINT] == 4011
+        assert usage_types[DataPointUsage.NO_CREATE] == 4389
 
-        assert len(ce_channels) == 133
+        assert len(ce_channels) == 135
         assert len(data_point_types) == 6
-        assert len(parameters) == 244
+        assert len(parameters) == 248
 
-        assert len(central.device_registry.devices) == 396
+        assert len(central.device_registry.devices) == 397
         virtual_remotes = ["VCU4264293", "VCU0000057", "VCU0000001"]
         await central.device_coordinator.delete_devices(interface_id=const.INTERFACE_ID, addresses=virtual_remotes)
-        assert len(central.device_registry.devices) == 393
+        assert len(central.device_registry.devices) == 394
         del_addresses = list(
             central.cache_coordinator.device_descriptions.get_device_descriptions(interface_id=const.INTERFACE_ID)
         )