+++ tests/test_model_data_point.py
@@ -28,7 +28,7 @@ class TestDataPointDefinition:
     def test_custom_required_data_points(self) -> None:
         """Test required parameters from data point definitions."""
         required_parameters = get_required_parameters()
-        assert len(required_parameters) == 100
+        assert len(required_parameters) == 103
         assert check_ignore_parameters_is_clean() is True